+++ Octokit.Reactive/Clients/IObservableAuthorizationsClient.cs
@@ -42,7 +42,7 @@ namespace Octokit.Reactive
         /// <returns>An <see cref="Authorization"/></returns>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
             Justification = "It's fiiiine. It's fine. Trust us.")]
-        IObservable<Authorization> Get(int id);
+        IObservable<Authorization> Get(long id);
 
         /// <summary>
         /// Creates a new personal token for the authenticated user.
@@ -219,7 +219,7 @@ namespace Octokit.Reactive
         /// <param name="id">The id of the <see cref="Authorization"/></param>
         /// <param name="authorizationUpdate">The changes to make to the authorization</param>
         /// <returns></returns>
-        IObservable<Authorization> Update(int id, AuthorizationUpdate authorizationUpdate);
+        IObservable<Authorization> Update(long id, AuthorizationUpdate authorizationUpdate);
 
         /// <summary>
         /// Deletes the specified <see cref="Authorization"/>.
@@ -234,7 +234,7 @@ namespace Octokit.Reactive
         /// Thrown when the current user does not have permission to make the request.
         /// </exception>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        IObservable<Unit> Delete(int id);
+        IObservable<Unit> Delete(long id);
 
         /// <summary>
         /// Deletes the specified <see cref="Authorization"/>.
@@ -250,6 +250,6 @@ namespace Octokit.Reactive
         /// Thrown when the current user does not have permission to make the request.
         /// </exception>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        IObservable<Unit> Delete(int id, string twoFactorAuthenticationCode);
+        IObservable<Unit> Delete(long id, string twoFactorAuthenticationCode);
     }
 }
